+++ Sources/Transifex/Core.swift
@@ -163,10 +163,11 @@ class NativeCore : TranslationProvider {
     ///   - translations: A list of `TXSourceString` objects.
     ///   - purge: Whether to replace the entire resource  content (true) or not (false). Defaults to false.
     ///   - completionHandler: A callback to be called when the push operation is complete with a
-    /// boolean argument that informs the caller that the operation was successful (true) or not (false).
+    /// boolean argument that informs the caller that the operation was successful (true) or not (false) and
+    /// an array that may or may not contain any errors produced during the push operation.
     func pushTranslations(_ translations: [TXSourceString],
                           purge: Bool = false,
-                          completionHandler: @escaping (Bool) -> Void) {
+                          completionHandler: @escaping (Bool, [Error]) -> Void) {
         cdsHandler.pushTranslations(translations,
                                     purge: purge,
                                     completionHandler: completionHandler)
@@ -323,7 +324,7 @@ render '\(stringToRender)' locale code: \(localeCode) params: \(params). Error:
 /// A static class that is the main point of entry for all the functionality of Transifex Native throughout the SDK.
 public final class TXNative : NSObject {
     /// The SDK version
-    internal static let version = "0.5.1"
+    internal static let version = "1.0.0"
     
     /// The filename of the file that holds the translated strings and it's bundled inside the app.
     public static let STRINGS_FILENAME = "txstrings.json"
@@ -471,11 +472,12 @@ token: \(token)
     ///   - translations: A list of `TXSourceString` objects.
     ///   - purge: Whether to replace the entire resource content (true) or not (false). Defaults to false.
     ///   - completionHandler: A callback to be called when the push operation is complete with a
-    /// boolean argument that informs the caller that the operation was successful (true) or not (false).
+    /// boolean argument that informs the caller that the operation was successful (true) or not (false) and
+    /// an array that may or may not contain any errors produced during the push operation.
     @objc
     public static func pushTranslations(_ translations: [TXSourceString],
                                         purge: Bool = false,
-                                        completionHandler: @escaping (Bool) -> Void) {
+                                        completionHandler: @escaping (Bool, [Error]) -> Void) {
         tx?.pushTranslations(translations,
                              purge: purge,
                              completionHandler: completionHandler)